refactoring files - working same
