+++ src_base/xeve_util.c
@@ -1991,60 +1991,6 @@ int xeve_check_cpu_info()
 }
 #endif
 
-XEVE_CTX * xeve_ctx_alloc(void)
-{
-    XEVE_CTX * ctx;
-
-    ctx = (XEVE_CTX*)xeve_malloc_fast(sizeof(XEVE_CTX));
-    xeve_assert_rv(ctx, NULL);
-    xeve_mset_x64a(ctx, 0, sizeof(XEVE_CTX));
-    return ctx;
-}
-
-void xeve_ctx_free(XEVE_CTX * ctx)
-{
-    xeve_mfree_fast(ctx);
-}
-
-
-XEVE_CORE * xeve_core_alloc(int chroma_format_idc)
-{
-    XEVE_CORE * core;
-    int i, j;
-
-    core = (XEVE_CORE *)xeve_malloc_fast(sizeof(XEVE_CORE));
-
-    xeve_assert_rv(core, NULL);
-    xeve_mset_x64a(core, 0, sizeof(XEVE_CORE));
-
-    for (i = 0; i < MAX_CU_LOG2; i++)
-    {
-        for (j = 0; j < MAX_CU_LOG2; j++)
-        {
-            xeve_create_cu_data(&core->cu_data_best[i][j], i, j, chroma_format_idc);
-            xeve_create_cu_data(&core->cu_data_temp[i][j], i, j, chroma_format_idc);
-        }
-    }
-
-    return core;
-}
-
-void xeve_core_free(XEVE_CORE * core)
-{
-    int i, j;
-
-    for (i = 0; i < MAX_CU_LOG2; i++)
-    {
-        for (j = 0; j < MAX_CU_LOG2; j++)
-        {
-            xeve_delete_cu_data(&core->cu_data_best[i][j], i, j);
-            xeve_delete_cu_data(&core->cu_data_temp[i][j], i, j);
-        }
-    }
-
-    xeve_mfree_fast(core);
-}
-
 void xeve_copy_chroma_qp_mapping_params(XEVE_CHROMA_TABLE *dst, XEVE_CHROMA_TABLE *src)
 {
     dst->chroma_qp_table_present_flag = src->chroma_qp_table_present_flag;
@@ -2054,159 +2000,6 @@ void xeve_copy_chroma_qp_mapping_params(XEVE_CHROMA_TABLE *dst, XEVE_CHROMA_TABL
     dst->num_points_in_qp_table_minus1[1] = src->num_points_in_qp_table_minus1[1];
     xeve_mcpy(&(dst->delta_qp_in_val_minus1), &(src->delta_qp_in_val_minus1), sizeof(int) * 2 * XEVE_MAX_QP_TABLE_SIZE);
     xeve_mcpy(&(dst->delta_qp_out_val), &(src->delta_qp_out_val), sizeof(int) * 2 * XEVE_MAX_QP_TABLE_SIZE);
-}
-
-static void parse_chroma_qp_mapping_table(XEVE_CHROMA_TABLE* chroma_qp_table, XEVE_PARAM * param)
-{
-    memset(chroma_qp_table, 0, sizeof(XEVE_CHROMA_TABLE));
-    chroma_qp_table->chroma_qp_table_present_flag = param->chroma_qp_table_present_flag;
-    if (chroma_qp_table->chroma_qp_table_present_flag)
-    {
-        chroma_qp_table->num_points_in_qp_table_minus1[0] = atoi(strtok(param->chroma_qp_num_points_in_table, " ")) - 1;
-        chroma_qp_table->num_points_in_qp_table_minus1[1] = atoi(strtok(NULL, " \r")) - 1;
-
-        { /* input pivot points */
-            chroma_qp_table->delta_qp_in_val_minus1[0][0] = atoi(strtok(param->chroma_qp_delta_in_val_cb, " "));
-            int j = 1;
-            do
-            {
-                char* val = strtok(NULL, " \r");
-                if (!val)
-                    break;
-                chroma_qp_table->delta_qp_in_val_minus1[0][j++] = atoi(val);
-            } while (1);
-            if (chroma_qp_table->num_points_in_qp_table_minus1[0] + 1 == j);
-
-            chroma_qp_table->delta_qp_in_val_minus1[1][0] = atoi(strtok(param->chroma_qp_delta_in_val_cr, " "));
-            j = 1;
-            do
-            {
-                char* val = strtok(NULL, " \r");
-                if (!val)
-                    break;
-                chroma_qp_table->delta_qp_in_val_minus1[1][j++] = atoi(val);
-            } while (1);
-            assert(chroma_qp_table->num_points_in_qp_table_minus1[1] + 1 == j);
-        }
-        {/* output pivot points */
-            chroma_qp_table->delta_qp_out_val[0][0] = atoi(strtok(param->chroma_qp_delta_out_val_cb, " "));
-            int j = 1;
-            do
-            {
-                char* val = strtok(NULL, " \r");
-                if (!val)
-                    break;
-                chroma_qp_table->delta_qp_out_val[0][j++] = atoi(val);
-            } while (1);
-            assert(chroma_qp_table->num_points_in_qp_table_minus1[0] + 1 == j);
-
-            chroma_qp_table->delta_qp_out_val[1][0] = atoi(strtok(param->chroma_qp_delta_out_val_cr, " "));
-            j = 1;
-            do
-            {
-                char* val = strtok(NULL, " \r");
-                if (!val)
-                    break;
-                chroma_qp_table->delta_qp_out_val[1][j++] = atoi(val);
-            } while (1);
-            assert(chroma_qp_table->num_points_in_qp_table_minus1[1] + 1 == j);
-        }
-    }
-}
-
-static void parse_chroma_qp_mapping_params(XEVE_CHROMA_TABLE *dst_struct, XEVE_CHROMA_TABLE *src_struct, int bit_depth)
-{
-    int qp_bd_offset_c = 6 * (bit_depth - 8);
-    dst_struct->chroma_qp_table_present_flag = src_struct->chroma_qp_table_present_flag;
-    dst_struct->num_points_in_qp_table_minus1[0] = src_struct->num_points_in_qp_table_minus1[0];
-    dst_struct->num_points_in_qp_table_minus1[1] = src_struct->num_points_in_qp_table_minus1[1];
-
-    if (dst_struct->chroma_qp_table_present_flag)
-    {
-        dst_struct->same_qp_table_for_chroma = 1;
-        if (src_struct->num_points_in_qp_table_minus1[0] != src_struct->num_points_in_qp_table_minus1[1])
-            dst_struct->same_qp_table_for_chroma = 0;
-        else
-        {
-            for (int i = 0; i < src_struct->num_points_in_qp_table_minus1[0]; i++)
-            {
-                if ((src_struct->delta_qp_in_val_minus1[0][i] != src_struct->delta_qp_in_val_minus1[1][i]) ||
-                    (src_struct->delta_qp_out_val[0][i] != src_struct->delta_qp_out_val[1][i]))
-                {
-                    dst_struct->same_qp_table_for_chroma = 0;
-                    break;
-                }
-            }
-        }
-
-        dst_struct->global_offset_flag = (src_struct->delta_qp_in_val_minus1[0][0] > 15 && src_struct->delta_qp_out_val[0][0] > 15) ? 1 : 0;
-        if (!dst_struct->same_qp_table_for_chroma)
-        {
-            dst_struct->global_offset_flag = dst_struct->global_offset_flag && ((src_struct->delta_qp_in_val_minus1[1][0] > 15 && src_struct->delta_qp_out_val[1][0] > 15) ? 1 : 0);
-        }
-
-        int start_qp = (dst_struct->global_offset_flag == 1) ? 16 : -qp_bd_offset_c;
-        for (int ch = 0; ch < (dst_struct->same_qp_table_for_chroma ? 1 : 2); ch++)
-        {
-            dst_struct->delta_qp_in_val_minus1[ch][0] = src_struct->delta_qp_in_val_minus1[ch][0] - start_qp;
-            dst_struct->delta_qp_out_val[ch][0] = src_struct->delta_qp_out_val[ch][0] - start_qp - dst_struct->delta_qp_in_val_minus1[ch][0];
-
-            for (int k = 1; k <= dst_struct->num_points_in_qp_table_minus1[ch]; k++)
-            {
-                dst_struct->delta_qp_in_val_minus1[ch][k] = (src_struct->delta_qp_in_val_minus1[ch][k] - src_struct->delta_qp_in_val_minus1[ch][k - 1]) - 1;
-                dst_struct->delta_qp_out_val[ch][k] = (src_struct->delta_qp_out_val[ch][k] - src_struct->delta_qp_out_val[ch][k - 1]) - (dst_struct->delta_qp_in_val_minus1[ch][k] + 1);
-            }
-        }
-    }
-}
-
-static void tbl_derived_chroma_qp_mapping(XEVE_CTX * ctx, XEVE_CHROMA_TABLE * struct_qp_c, int bit_depth)
-{
-    int MAX_QP = XEVE_MAX_QP_TABLE_SIZE - 1;
-    int qpInVal[XEVE_MAX_QP_TABLE_SIZE_EXT] = { 0 };
-    int qpOutVal[XEVE_MAX_QP_TABLE_SIZE_EXT] = { 0 };
-    int qp_bd_offset_c = 6 * (bit_depth - 8);
-    int startQp = (struct_qp_c->global_offset_flag == 1) ? 16 : -qp_bd_offset_c;
-
-    for (int i = 0; i < (struct_qp_c->same_qp_table_for_chroma ? 1 : 2); i++)
-    {
-        qpInVal[0] = startQp + struct_qp_c->delta_qp_in_val_minus1[i][0];
-        qpOutVal[0] = startQp + struct_qp_c->delta_qp_in_val_minus1[i][0] + struct_qp_c->delta_qp_out_val[i][0];
-        for (int j = 1; j <= struct_qp_c->num_points_in_qp_table_minus1[i]; j++)
-        {
-            qpInVal[j] = qpInVal[j - 1] + struct_qp_c->delta_qp_in_val_minus1[i][j] + 1;
-            qpOutVal[j] = qpOutVal[j - 1] + (struct_qp_c->delta_qp_in_val_minus1[i][j] + 1 + struct_qp_c->delta_qp_out_val[i][j]);
-        }
-
-        for (int j = 0; j <= struct_qp_c->num_points_in_qp_table_minus1[i]; j++)
-        {
-            assert(qpInVal[j] >= -qp_bd_offset_c && qpInVal[j]  <= MAX_QP);
-            assert(qpOutVal[j] >= -qp_bd_offset_c && qpOutVal[j] <= MAX_QP);
-        }
-
-        ctx->qp_chroma_dynamic[i][qpInVal[0]] = qpOutVal[0];
-        for (int k = qpInVal[0] - 1; k >= -qp_bd_offset_c; k--)
-        {
-            ctx->qp_chroma_dynamic[i][k] = XEVE_CLIP3(-qp_bd_offset_c, MAX_QP, ctx->qp_chroma_dynamic[i][k + 1] - 1);
-        }
-        for (int j = 0; j < struct_qp_c->num_points_in_qp_table_minus1[i]; j++)
-        {
-            int sh = (struct_qp_c->delta_qp_in_val_minus1[i][j + 1] + 1) >> 1;
-            for (int k = qpInVal[j] + 1, m = 1; k <= qpInVal[j + 1]; k++, m++)
-            {
-                ctx->qp_chroma_dynamic[i][k] = ctx->qp_chroma_dynamic[i][qpInVal[j]]
-                    + ((qpOutVal[j + 1] - qpOutVal[j]) * m + sh) / (struct_qp_c->delta_qp_in_val_minus1[i][j + 1] + 1);
-            }
-        }
-        for (int k = qpInVal[struct_qp_c->num_points_in_qp_table_minus1[i]] + 1; k <= MAX_QP; k++)
-        {
-            ctx->qp_chroma_dynamic[i][k] = XEVE_CLIP3(-qp_bd_offset_c, MAX_QP, ctx->qp_chroma_dynamic[i][k - 1] + 1);
-        }
-    }
-    if (struct_qp_c->same_qp_table_for_chroma)
-    {
-        xeve_mcpy(&(ctx->qp_chroma_dynamic[1][-qp_bd_offset_c]), &(ctx->qp_chroma_dynamic[0][-qp_bd_offset_c]), XEVE_MAX_QP_TABLE_SIZE_EXT * sizeof(int));
-    }
 }
 
 // ChromaQP offset for U and V components
@@ -2219,1186 +2012,6 @@ void xeve_set_chroma_qp_tbl_loc(XEVE_CTX* ctx)
     }
     ctx->qp_chroma_dynamic[0] = &(ctx->qp_chroma_dynamic_ext[0][6 * (ctx->param.codec_bit_depth - 8)]);
     ctx->qp_chroma_dynamic[1] = &(ctx->qp_chroma_dynamic_ext[1][6 * (ctx->param.codec_bit_depth - 8)]);
-}
-
-static int parse_tile_slice_param(XEVE_CTX* ctx)
-{
-    XEVE_TS_INFO * ts_info = &ctx->ts_info;
-    XEVE_PARAM   * param = &ctx->param;
-
-    ts_info->tile_uniform_spacing_flag = param->tile_uniform_spacing_flag;
-    ts_info->tile_columns              = param->tile_columns;
-    ts_info->tile_rows                 = param->tile_rows;
-    ts_info->num_slice_in_pic          = param->num_slice_in_pic;
-    ts_info->arbitrary_slice_flag      = param->arbitrary_slice_flag;
-
-    int num_tiles = ts_info->tile_columns * ts_info->tile_rows;
-    if (num_tiles < ts_info->num_slice_in_pic) return XEVE_ERR;
-    if (num_tiles > 1)
-    {
-        if (!ts_info->tile_uniform_spacing_flag)
-        {
-            ts_info->tile_column_width_array[0] = atoi(strtok(param->tile_column_width_array, " "));
-            int j = 1;
-            do
-            {
-                char* val = strtok(NULL, " \r");
-                if (!val)
-                    break;
-                ts_info->tile_column_width_array[j++] = atoi(val);
-            } while (1);
-
-            ts_info->tile_row_height_array[0] = atoi(strtok(param->tile_row_height_array, " "));
-            j = 1;
-            do
-            {
-                char* val = strtok(NULL, " \r");
-                if (!val)
-                    break;
-                ts_info->tile_row_height_array[j++] = atoi(val);
-            } while (1);
-        }
-
-        if (ts_info->num_slice_in_pic == 1)
-        {
-            ts_info->tile_array_in_slice[0] = 0;
-            ts_info->tile_array_in_slice[1] = (ts_info->tile_columns * ts_info->tile_rows) - 1;
-            ts_info->num_remaining_tiles_in_slice_minus1[0] = param->num_remaining_tiles_in_slice_minus1[0];
-        }
-        else /* There are more than one slice in the picture */
-        {
-            ts_info->tile_array_in_slice[0] = atoi(strtok(param->num_remaining_tiles_in_slice_minus1, " "));
-            int j = 1;
-            do
-            {
-                char* val = strtok(NULL, " \r");
-                if (!val)
-                    break;
-                ts_info->tile_array_in_slice[j++] = atoi(val);
-            } while (1);
-
-            if (ts_info->arbitrary_slice_flag)
-            {
-                ts_info->num_remaining_tiles_in_slice_minus1[0] = atoi(strtok(param->num_remaining_tiles_in_slice_minus1, " ")) - 1;
-                int j = 1;
-                do
-                {
-                    char* val = strtok(NULL, " \r");
-                    if (!val)
-                        break;
-                    ts_info->num_remaining_tiles_in_slice_minus1[j++] = atoi(val) - 1;
-                } while (1);
-            }
-        }
-    }
-
-    return XEVE_OK;
-}
-
-int xeve_set_init_param(XEVE_CTX * ctx, XEVE_PARAM * param)
-{
-    /* check input parameters */
-    xeve_assert_rv(param->w > 0 && param->h > 0, XEVE_ERR_INVALID_ARGUMENT);
-    xeve_assert_rv(param->qp >= MIN_QUANT && param->qp <= MAX_QUANT, XEVE_ERR_INVALID_ARGUMENT);
-    xeve_assert_rv(param->keyint >= 0 ,XEVE_ERR_INVALID_ARGUMENT);
-    xeve_assert_rv(param->threads <= XEVE_MAX_THREADS ,XEVE_ERR_INVALID_ARGUMENT);
-
-    if(param->disable_hgop == 0)
-    {
-        xeve_assert_rv(param->bframes == 0 || param->bframes == 1 || \
-                       param->bframes == 3 || param->bframes == 7 || \
-                       param->bframes == 15, XEVE_ERR_INVALID_ARGUMENT);
-
-        if(param->bframes != 0)
-        {
-            if(param->keyint % (param->bframes + 1) != 0)
-            {
-                xeve_assert_rv(0, XEVE_ERR_INVALID_ARGUMENT);
-            }
-        }
-    }
-
-    if (param->ref != 0)
-    {
-        if (param->bframes == 0)
-        {
-            param->me_ref_num = XEVE_MIN(5, param->ref);
-        }
-        else
-        {
-            param->me_ref_num = (param->ref > param->bframes) ? param->bframes : param->ref;
-        }
-    }
-
-    if (param->ref_pic_gap_length != 0)
-    {
-        xeve_assert_rv(param->bframes == 0, XEVE_ERR_INVALID_ARGUMENT);
-    }
-
-    if (param->bframes == 0)
-    {
-        if (param->ref_pic_gap_length == 0)
-        {
-            param->ref_pic_gap_length = 1;
-        }
-        xeve_assert_rv(param->ref_pic_gap_length == 1 || param->ref_pic_gap_length == 2 || \
-                       param->ref_pic_gap_length == 4 || param->ref_pic_gap_length == 8 || \
-                       param->ref_pic_gap_length == 16, XEVE_ERR_INVALID_ARGUMENT);
-    }
-
-    /* set default encoding parameter */
-    param->qp_max            = MAX_QUANT;
-    param->qp_min            = MIN_QUANT;
-    param->gop_size          = param->bframes +1;
-    param->lookahead         = XEVE_MIN(XEVE_MAX((param->cutree)? param->gop_size : 0, param->lookahead), XEVE_MAX_INBUF_CNT>>1);
-    param->use_fcst          = ((param->use_fcst || param->lookahead) && (param->rc_type || param->aq_mode)) ? 1 : 0;
-    param->chroma_format_idc = XEVE_CFI_FROM_CF(XEVE_CS_GET_FORMAT(param->cs));
-    param->cs_w_shift        = XEVE_GET_CHROMA_W_SHIFT(param->chroma_format_idc);
-    param->cs_h_shift        = XEVE_GET_CHROMA_H_SHIFT(param->chroma_format_idc);
-
-
-    if (param->chroma_qp_table_present_flag)
-    {
-        XEVE_CHROMA_TABLE tmp_qp_tbl;
-        parse_chroma_qp_mapping_table(&tmp_qp_tbl, param);
-        parse_chroma_qp_mapping_params(&(ctx->chroma_qp_table_struct), &tmp_qp_tbl, param->codec_bit_depth);
-        tbl_derived_chroma_qp_mapping(ctx, &(ctx->chroma_qp_table_struct), param->codec_bit_depth);
-    }
-    else
-    {
-        const int * qp_chroma_ajudst = xeve_tbl_qp_chroma_ajudst;
-        xeve_mcpy(&(ctx->qp_chroma_dynamic_ext[0][6 *( param->codec_bit_depth - 8)]), qp_chroma_ajudst, XEVE_MAX_QP_TABLE_SIZE * sizeof(int));
-        xeve_mcpy(&(ctx->qp_chroma_dynamic_ext[1][6 * (param->codec_bit_depth - 8)]), qp_chroma_ajudst, XEVE_MAX_QP_TABLE_SIZE * sizeof(int));
-    }
-
-#if 0
-    if (param->vbv_bufsize == 0)
-    {
-        param->vbv_bufsize = (int)((param->bitrate) * (param->vbv_msec / 1000.0));
-    }
-#endif
-
-    parse_tile_slice_param(ctx);
-
-    return XEVE_OK;
-}
-
-static void decide_normal_gop(XEVE_CTX * ctx, u32 pic_imcnt)
-{
-    int i_period, gop_size, pos;
-    u32        pic_icnt_b;
-
-    i_period = ctx->param.keyint;
-    gop_size = ctx->param.gop_size;
-
-    if (i_period == 0 && pic_imcnt == 0)
-    {
-        ctx->slice_type = SLICE_I;
-        ctx->slice_depth = FRM_DEPTH_0;
-        ctx->poc.poc_val = pic_imcnt;
-        ctx->poc.prev_doc_offset = 0;
-        ctx->poc.prev_poc_val = ctx->poc.poc_val;
-        ctx->slice_ref_flag = 1;
-    }
-    else if ((i_period != 0) && pic_imcnt % i_period == 0)
-    {
-        ctx->slice_type = SLICE_I;
-        ctx->slice_depth = FRM_DEPTH_0;
-        ctx->poc.poc_val = pic_imcnt;
-        ctx->poc.prev_doc_offset = 0;
-        ctx->poc.prev_poc_val = ctx->poc.poc_val;
-        ctx->slice_ref_flag = 1;
-    }
-    else if (pic_imcnt % gop_size == 0)
-    {
-        ctx->slice_type = ctx->param.inter_slice_type;
-        ctx->slice_ref_flag = 1;
-        ctx->slice_depth = FRM_DEPTH_1;
-        ctx->poc.poc_val = pic_imcnt;
-        ctx->poc.prev_doc_offset = 0;
-        ctx->poc.prev_poc_val = ctx->poc.poc_val;
-        ctx->slice_ref_flag = 1;
-    }
-    else
-    {
-        ctx->slice_type = ctx->param.inter_slice_type;
-        if (ctx->param.disable_hgop == 0)
-        {
-            pos = (pic_imcnt % gop_size) - 1;
-
-            if (ctx->sps.tool_pocs)
-            {
-                ctx->fn_pocs(ctx, pic_imcnt, gop_size, pos);
-            }
-            else
-            {
-                ctx->slice_depth = xeve_tbl_slice_depth[gop_size >> 2][pos];
-                int tid = ctx->slice_depth - (ctx->slice_depth > 0);
-                xeve_poc_derivation(ctx->sps, tid, &ctx->poc);
-            }
-            if (!ctx->sps.tool_pocs && gop_size >= 2)
-            {
-                ctx->slice_ref_flag = (ctx->slice_depth == xeve_tbl_slice_depth[gop_size >> 2][gop_size - 2] ? 0 : 1);
-            }
-            else
-            {
-                ctx->slice_ref_flag = 1;
-            }
-        }
-        else
-        {
-            pos = (pic_imcnt % gop_size) - 1;
-            ctx->slice_depth = FRM_DEPTH_2;
-            ctx->poc.poc_val = ((pic_imcnt / gop_size) * gop_size) - gop_size + pos + 1;
-            ctx->slice_ref_flag = 0;
-        }
-        /* find current encoding picture's(B picture) pic_icnt */
-        pic_icnt_b = ctx->poc.poc_val;
-
-        /* find pico again here */
-        ctx->pico_idx = (u8)(pic_icnt_b % ctx->pico_max_cnt);
-        ctx->pico = ctx->pico_buf[ctx->pico_idx];
-
-        PIC_ORIG(ctx) = &ctx->pico->pic;
-    }
-}
-
-
-/* slice_type / slice_depth / poc / PIC_ORIG setting */
-static void decide_slice_type(XEVE_CTX * ctx)
-{
-    u32 pic_imcnt, pic_icnt;
-    int i_period, gop_size;
-    int force_cnt = 0;
-
-    i_period = ctx->param.keyint;
-    gop_size = ctx->param.gop_size;
-    pic_icnt = (ctx->pic_cnt + ctx->param.bframes);
-    pic_imcnt = pic_icnt;
-    ctx->pico_idx = pic_icnt % ctx->pico_max_cnt;
-    ctx->pico = ctx->pico_buf[ctx->pico_idx];
-    PIC_ORIG(ctx) = &ctx->pico->pic;
-
-    if(gop_size == 1)
-    {
-        if (i_period == 1) /* IIII... */
-        {
-            ctx->slice_type = SLICE_I;
-            ctx->slice_depth = FRM_DEPTH_0;
-            ctx->poc.poc_val = pic_icnt;
-            ctx->slice_ref_flag = 0;
-        }
-        else /* IPPP... */
-        {
-            pic_imcnt = (i_period > 0) ? pic_icnt % i_period : pic_icnt;
-            if (pic_imcnt == 0)
-            {
-                ctx->slice_type = SLICE_I;
-                ctx->slice_depth = FRM_DEPTH_0;
-                ctx->slice_ref_flag = 1;
-            }
-            else
-            {
-                ctx->slice_type = ctx->param.inter_slice_type;
-
-                if (ctx->param.disable_hgop == 0)
-                {
-                    ctx->slice_depth = xeve_tbl_slice_depth_P[ctx->param.ref_pic_gap_length >> 2][(pic_imcnt - 1) % ctx->param.ref_pic_gap_length];
-                }
-                else
-                {
-                    ctx->slice_depth = FRM_DEPTH_1;
-                }
-                ctx->slice_ref_flag = 1;
-            }
-            ctx->poc.poc_val = (ctx->param.closed_gop && i_period > 0 && (ctx->pic_cnt % i_period) == 0 ?
-                                0 : (ctx->param.closed_gop ? ctx->pic_cnt % i_period : ctx->pic_cnt));
-        }
-    }
-    else /* include B Picture (gop_size = 2 or 4 or 8 or 16) */
-    {
-        if(pic_icnt == gop_size - 1) /* special case when sequence start */
-        {
-            ctx->slice_type = SLICE_I;
-            ctx->slice_depth = FRM_DEPTH_0;
-            ctx->poc.poc_val = 0;
-            ctx->poc.prev_doc_offset = 0;
-            ctx->poc.prev_poc_val = ctx->poc.poc_val;
-            ctx->slice_ref_flag = 1;
-
-            /* flush the first IDR picture */
-            PIC_ORIG(ctx) = &ctx->pico_buf[0]->pic;
-            ctx->pico = ctx->pico_buf[0];
-        }
-        else if(ctx->force_slice)
-        {
-            for(force_cnt = ctx->force_ignored_cnt; force_cnt < gop_size; force_cnt++)
-            {
-                pic_icnt = (ctx->pic_cnt + ctx->param.bframes + force_cnt);
-                pic_imcnt = pic_icnt;
-
-                decide_normal_gop(ctx, pic_imcnt);
-
-                if(ctx->poc.poc_val <= (int)ctx->pic_ticnt)
-                {
-                    break;
-                }
-            }
-            ctx->force_ignored_cnt = force_cnt;
-        }
-        else /* normal GOP case */
-        {
-            decide_normal_gop(ctx, pic_imcnt);
-        }
-    }
-    if (ctx->param.disable_hgop == 0 && gop_size > 1)
-    {
-        ctx->nalu.nuh_temporal_id = ctx->slice_depth - (ctx->slice_depth > 0);
-    }
-    else
-    {
-        ctx->nalu.nuh_temporal_id = 0;
-    }
-    if (ctx->slice_type == SLICE_I && ctx->param.closed_gop)
-    {
-        ctx->poc.prev_idr_poc = ctx->poc.poc_val;
-    }
-}
-
-
-int xeve_pic_prepare(XEVE_CTX * ctx, XEVE_BITB * bitb, XEVE_STAT * stat)
-{
-    int             ret;
-    int             size;
-
-    xeve_assert_rv(PIC_ORIG(ctx) != NULL, XEVE_ERR_UNEXPECTED);
-
-    ctx->qp = (u8)ctx->param.qp;
-
-    PIC_CURR(ctx) = xeve_picman_get_empty_pic(&ctx->rpm, &ret);
-    xeve_assert_rv(PIC_CURR(ctx) != NULL, ret);
-    ctx->map_refi = PIC_CURR(ctx)->map_refi;
-    ctx->map_mv = PIC_CURR(ctx)->map_mv;
-    ctx->map_unrefined_mv = PIC_CURR(ctx)->map_unrefined_mv;
-    ctx->map_dqp_lah = ctx->pico->sinfo.map_qp_scu;
-
-    PIC_MODE(ctx) = PIC_CURR(ctx);
-    if(ctx->pic_dbk == NULL)
-    {
-        ctx->pic_dbk = xeve_pic_alloc(&ctx->rpm.pa, &ret);
-        xeve_assert_rv(ctx->pic_dbk != NULL, ret);
-    }
-
-    decide_slice_type(ctx);
-
-    ctx->lcu_cnt = ctx->f_lcu;
-    ctx->slice_num = 0;
-
-    if (ctx->tile_cnt == 1 && ctx->param.threads > 1)
-    {
-        for (u32 i = 0; i < ctx->f_lcu; i++)
-        {
-            ctx->sync_flag[i] = 0; //Reset the sync flag at the begining of each frame
-        }
-    }
-
-    if(ctx->slice_type == SLICE_I) ctx->last_intra_poc = ctx->poc.poc_val;
-
-    size = sizeof(s8) * ctx->f_scu * REFP_NUM;
-    xeve_mset_x64a(ctx->map_refi, -1, size);
-    size = sizeof(s16) * ctx->f_scu * REFP_NUM * MV_D;
-    xeve_mset_x64a(ctx->map_mv, 0, size);
-    size = sizeof(s16) * ctx->f_scu * REFP_NUM * MV_D;
-    xeve_mset_x64a(ctx->map_unrefined_mv, 0, size);
-
-    /* initialize bitstream container */
-    xeve_bsw_init(&ctx->bs[0], bitb->addr, bitb->bsize, NULL);
-    ctx->bs[0].pdata[1] = &ctx->sbac_enc[0];
-    for (int i = 0; i < ctx->param.threads; i++)
-    {
-        xeve_bsw_init(&ctx->bs[i], ctx->bs[i].beg, bitb->bsize, NULL);
-    }
-
-    /* clear map */
-    xeve_mset_x64a(ctx->map_scu, 0, sizeof(u32) * ctx->f_scu);
-    xeve_mset_x64a(ctx->map_cu_mode, 0, sizeof(u32) * ctx->f_scu);
-
-    xeve_set_active_pps_info(ctx);
-    if (ctx->param.rc_type != 0)
-    {
-        ctx->qp = xeve_rc_get_qp(ctx);
-    }
-
-    return XEVE_OK;
-}
-
-int xeve_pic_finish(XEVE_CTX *ctx, XEVE_BITB *bitb, XEVE_STAT *stat)
-{
-    XEVE_IMGB *imgb_o, *imgb_c;
-    int        ret;
-    int        i, j;
-
-    xeve_mset(stat, 0, sizeof(XEVE_STAT));
-
-    /* adding picture sign */
-    if (ctx->param.use_pic_sign)
-    {
-        XEVE_BSW * bs = &ctx->bs[0];
-        XEVE_NALU sei_nalu;
-        xeve_set_nalu(&sei_nalu, XEVE_SEI_NUT, ctx->nalu.nuh_temporal_id);
-
-        u8* size_field = bs->cur;
-        u8* cur_tmp = bs->cur;
-
-        xeve_eco_nalu(bs, &sei_nalu);
-
-        ret = xeve_eco_sei(ctx, bs);
-        xeve_assert_rv(ret == XEVE_OK, ret);
-
-        xeve_bsw_deinit(bs);
-        stat->sei_size = (int)(bs->cur - cur_tmp);
-        xeve_eco_nal_unit_len(size_field, stat->sei_size - 4);
-    }
-
-    /* expand current encoding picture, if needs */
-    ctx->fn_picbuf_expand(ctx, PIC_CURR(ctx));
-
-    /* picture buffer management */
-    ret = xeve_picman_put_pic(&ctx->rpm, PIC_CURR(ctx), ctx->nalu.nal_unit_type_plus1 - 1 == XEVE_IDR_NUT,
-                              ctx->poc.poc_val, ctx->nalu.nuh_temporal_id, 0, ctx->refp,
-                              ctx->slice_ref_flag, ctx->sps.tool_rpl, ctx->param.ref_pic_gap_length);
-
-    xeve_assert_rv(ret == XEVE_OK, ret);
-
-    imgb_o = PIC_ORIG(ctx)->imgb;
-    xeve_assert(imgb_o != NULL);
-
-    imgb_c = PIC_CURR(ctx)->imgb;
-    xeve_assert(imgb_c != NULL);
-
-    /* set stat */
-    stat->write = XEVE_BSW_GET_WRITE_BYTE(&ctx->bs[0]);
-    stat->nalu_type = (ctx->slice_type == SLICE_I && ctx->param.closed_gop) ? XEVE_IDR_NUT : XEVE_NONIDR_NUT;
-    stat->stype = ctx->slice_type;
-    stat->fnum = ctx->pic_cnt;
-    stat->qp = ctx->sh->qp;
-    stat->poc = ctx->poc.poc_val;
-    stat->tid = ctx->nalu.nuh_temporal_id;
-
-    for(i = 0; i < 2; i++)
-    {
-        stat->refpic_num[i] = ctx->rpm.num_refp[i];
-        for (j = 0; j < stat->refpic_num[i]; j++)
-        {
-            stat->refpic[i][j] = ctx->refp[j][i].poc;
-        }
-    }
-
-    ctx->pic_cnt++; /* increase picture count */
-    ctx->param.f_ifrm = 0; /* clear force-IDR flag */
-    ctx->pico->is_used = 0;
-
-    imgb_c->ts[0] = bitb->ts[0] = imgb_o->ts[0];
-    imgb_c->ts[1] = bitb->ts[1] = imgb_o->ts[1];
-    imgb_c->ts[2] = bitb->ts[2] = imgb_o->ts[2];
-    imgb_c->ts[3] = bitb->ts[3] = imgb_o->ts[3];
-
-    if (ctx->param.rc_type != 0)
-    {
-        ctx->rcore->real_bits = (stat->write - stat->sei_size) << 3;
-    }
-
-    if (imgb_o)
-    {
-        imgb_o->release(imgb_o);
-    }
-
-    return XEVE_OK;
-}
-
-void xeve_set_nalu(XEVE_NALU * nalu, int nalu_type, int nuh_temporal_id)
-{
-    nalu->nal_unit_size = 0;
-    nalu->forbidden_zero_bit = 0;
-    nalu->nal_unit_type_plus1 = nalu_type + 1;
-    nalu->nuh_temporal_id = nuh_temporal_id;
-    nalu->nuh_reserved_zero_5bits = 0;
-    nalu->nuh_extension_flag = 0;
-}
-
-void xeve_set_vui(XEVE_CTX * ctx, XEVE_VUI * vui)
-{
-    vui->aspect_ratio_info_present_flag = 1;
-    vui->aspect_ratio_idc = 1;
-    vui->sar_width = 1;
-    vui->sar_height = 1;
-    vui->overscan_info_present_flag = 1;
-    vui->overscan_appropriate_flag = 1;
-    vui->video_signal_type_present_flag = 1;
-    vui->video_format = 1;
-    vui->video_full_range_flag = 1;
-    vui->colour_description_present_flag = 1;
-    vui->colour_primaries = 1;
-    vui->transfer_characteristics = 1;
-    vui->matrix_coefficients = 1;
-    vui->chroma_loc_info_present_flag = 1;
-    vui->chroma_sample_loc_type_top_field = 1;
-    vui->chroma_sample_loc_type_bottom_field = 1;
-    vui->neutral_chroma_indication_flag = 1;
-    vui->field_seq_flag = 1;
-    vui->timing_info_present_flag = 1;
-    vui->num_units_in_tick = 1;
-    vui->time_scale = 1;
-    vui->fixed_pic_rate_flag = 1;
-    vui->nal_hrd_parameters_present_flag = 1;
-    vui->vcl_hrd_parameters_present_flag = 1;
-    vui->low_delay_hrd_flag = 1;
-    vui->pic_struct_present_flag = 1;
-    vui->bitstream_restriction_flag = 1;
-    vui->motion_vectors_over_pic_boundaries_flag = 1;
-    vui->max_bytes_per_pic_denom = 1;
-    vui->max_bits_per_mb_denom = 1;
-    vui->log2_max_mv_length_horizontal = 1;
-    vui->log2_max_mv_length_vertical = 1;
-    vui->num_reorder_pics = 1;
-    vui->max_dec_pic_buffering = 1;
-    vui->hrd_parameters.cpb_cnt_minus1 = 1;
-    vui->hrd_parameters.bit_rate_scale = 1;
-    vui->hrd_parameters.cpb_size_scale = 1;
-    xeve_mset(&(vui->hrd_parameters.bit_rate_value_minus1), 0, sizeof(int)*NUM_CPB);
-    xeve_mset(&(vui->hrd_parameters.cpb_size_value_minus1), 0, sizeof(int)*NUM_CPB);
-    xeve_mset(&(vui->hrd_parameters.cbr_flag), 0, sizeof(int)*NUM_CPB);
-    vui->hrd_parameters.initial_cpb_removal_delay_length_minus1 = 1;
-    vui->hrd_parameters.cpb_removal_delay_length_minus1 = 1;
-    vui->hrd_parameters.dpb_output_delay_length_minus1 = 1;
-    vui->hrd_parameters.time_offset_length = 1;
-}
-
-void xeve_set_sps(XEVE_CTX * ctx, XEVE_SPS * sps)
-{
-    xeve_mset(sps, 0, sizeof(XEVE_SPS));
-
-    sps->profile_idc = ctx->param.profile;
-    sps->level_idc = ctx->param.level_idc * 3;
-    sps->pic_width_in_luma_samples = XEVE_ALIGN_VAL(ctx->param.w, 8);
-    sps->pic_height_in_luma_samples = XEVE_ALIGN_VAL(ctx->param.h, 8);
-    sps->toolset_idc_h = 0;
-    sps->toolset_idc_l = 0;
-    sps->bit_depth_luma_minus8 = ctx->param.codec_bit_depth - 8;
-    sps->bit_depth_chroma_minus8 = ctx->param.codec_bit_depth - 8;
-    sps->chroma_format_idc = XEVE_CFI_FROM_CF(XEVE_CS_GET_FORMAT(ctx->param.cs));
-    sps->dquant_flag = 0;
-    sps->log2_max_pic_order_cnt_lsb_minus4 = POC_LSB_BIT - 4;
-    sps->sps_max_dec_pic_buffering_minus1 = 0; //[TBF]
-
-    if (ctx->param.bframes > 0)
-    {
-        sps->max_num_ref_pics = ctx->param.me_ref_num;
-    }
-    else
-    {
-        sps->max_num_ref_pics = ctx->param.ref_pic_gap_length;;
-    }
-
-    sps->log2_sub_gop_length = (int)(log2(ctx->param.gop_size) + .5);
-    sps->log2_ref_pic_gap_length = (int)(log2(ctx->param.ref_pic_gap_length) + .5);
-    sps->long_term_ref_pics_flag = 0;
-    sps->vui_parameters_present_flag = 0;
-    xeve_set_vui(ctx, &(sps->vui_parameters));
-
-    if (ctx->chroma_qp_table_struct.chroma_qp_table_present_flag)
-    {
-        xeve_copy_chroma_qp_mapping_params(&(sps->chroma_qp_table_struct), &(ctx->chroma_qp_table_struct));
-    }
-
-    sps->picture_cropping_flag = ctx->param.picture_cropping_flag;
-    if (sps->picture_cropping_flag)
-    {
-        sps->picture_crop_left_offset = ctx->param.picture_crop_left_offset;
-        sps->picture_crop_right_offset = ctx->param.picture_crop_right_offset;
-        sps->picture_crop_top_offset = ctx->param.picture_crop_top_offset;
-        sps->picture_crop_bottom_offset = ctx->param.picture_crop_bottom_offset;
-    }
-}
-
-int xeve_set_active_pps_info(XEVE_CTX * ctx)
-{
-    int active_pps_id = ctx->sh->slice_pic_parameter_set_id;
-    xeve_mcpy(&(ctx->pps), &(ctx->pps_array[active_pps_id]), sizeof(XEVE_PPS));
-
-    return XEVE_OK;
-}
-
-void xeve_set_pps(XEVE_CTX * ctx, XEVE_PPS * pps)
-{
-    pps->loop_filter_across_tiles_enabled_flag = 0;
-    pps->single_tile_in_pic_flag = 1;
-    pps->constrained_intra_pred_flag = ctx->param.constrained_intra_pred;
-    pps->cu_qp_delta_enabled_flag = (ctx->param.aq_mode || ctx->param.cutree);
-
-    pps->num_ref_idx_default_active_minus1[REFP_0] = 0;
-    pps->num_ref_idx_default_active_minus1[REFP_1] = 0;
-
-    ctx->pps.pps_pic_parameter_set_id = 0;
-    xeve_mcpy(&ctx->pps_array[ctx->pps.pps_pic_parameter_set_id], &ctx->pps, sizeof(XEVE_PPS));
-}
-
-void xeve_set_sh(XEVE_CTX *ctx, XEVE_SH *sh)
-{
-    double qp;
-    int qp_l_i;
-    int qp_c_i;
-
-    const QP_ADAPT_PARAM *qp_adapt_param = ctx->param.bframes == 0 ?
-        (ctx->param.keyint == 1 ? xeve_qp_adapt_param_ai : xeve_qp_adapt_param_ld) : xeve_qp_adapt_param_ra;
-
-    sh->slice_type = ctx->slice_type;
-    sh->no_output_of_prior_pics_flag = 0;
-    sh->deblocking_filter_on = (ctx->param.use_deblock) ? 1 : 0;
-    sh->sh_deblock_alpha_offset = ctx->param.deblock_alpha_offset;
-    sh->sh_deblock_beta_offset = ctx->param.deblock_beta_offset;
-    sh->single_tile_in_slice_flag = 1;
-    sh->collocated_from_list_idx = (sh->slice_type == SLICE_P) ? REFP_0 : REFP_1;  // Specifies source (List ID) of the collocated picture, equialent of the collocated_from_l0_flag
-    sh->collocated_from_ref_idx = 0;        // Specifies source (RefID_ of the collocated picture, equialent of the collocated_ref_idx
-    sh->collocated_mvp_source_list_idx = REFP_0;  // Specifies source (List ID) in collocated pic that provides MV information (Applicability is function of NoBackwardPredFlag)
-
-    /* set lambda */
-    qp = XEVE_CLIP3(0, MAX_QUANT, (ctx->param.qp_incread_frame != 0 && (int)(ctx->poc.poc_val) >= ctx->param.qp_incread_frame) ? ctx->qp + 1.0 : ctx->qp);
-    sh->dqp = ctx->param.aq_mode != 0;
-
-    if(ctx->param.disable_hgop == 0 && ctx->param.rc_type == 0)
-    {
-        double dqp_offset;
-        int qp_offset;
-
-        qp += qp_adapt_param[ctx->slice_depth].qp_offset_layer;
-        dqp_offset = qp * qp_adapt_param[ctx->slice_depth].qp_offset_model_scale + qp_adapt_param[ctx->slice_depth].qp_offset_model_offset + 0.5;
-
-        qp_offset = (int)floor(XEVE_CLIP3(0.0, 3.0, dqp_offset));
-        qp += qp_offset;
-    }
-
-    sh->qp   = (u8)XEVE_CLIP3(0, MAX_QUANT, qp);
-    sh->qp_u_offset = ctx->param.qp_cb_offset;
-    sh->qp_v_offset = ctx->param.qp_cr_offset;
-    sh->qp_u = (s8)XEVE_CLIP3(-6 * ctx->sps.bit_depth_chroma_minus8, 57, sh->qp + sh->qp_u_offset);
-    sh->qp_v = (s8)XEVE_CLIP3(-6 * ctx->sps.bit_depth_chroma_minus8, 57, sh->qp + sh->qp_v_offset);
-
-    qp_l_i = sh->qp;
-    ctx->lambda[0] = 0.57 * pow(2.0, (qp_l_i - 12.0) / 3.0);
-    qp_c_i = ctx->qp_chroma_dynamic[0][sh->qp_u];
-    ctx->dist_chroma_weight[0] = pow(2.0, (qp_l_i - qp_c_i) / 3.0);
-    qp_c_i = ctx->qp_chroma_dynamic[1][sh->qp_v];
-    ctx->dist_chroma_weight[1] = pow(2.0, (qp_l_i - qp_c_i) / 3.0);
-    ctx->lambda[1] = ctx->lambda[0] / ctx->dist_chroma_weight[0];
-    ctx->lambda[2] = ctx->lambda[0] / ctx->dist_chroma_weight[1];
-    ctx->sqrt_lambda[0] = sqrt(ctx->lambda[0]);
-    ctx->sqrt_lambda[1] = sqrt(ctx->lambda[1]);
-    ctx->sqrt_lambda[2] = sqrt(ctx->lambda[2]);
-
-    ctx->sh->slice_pic_parameter_set_id = 0;
-}
-
-int xeve_set_tile_info(XEVE_CTX * ctx)
-{
-    XEVE_TILE  * tile;
-    int          size, f_tile, tidx;
-
-    ctx->tile_cnt = ctx->ts_info.tile_columns * ctx->ts_info.tile_rows;
-    f_tile = ctx->ts_info.tile_columns * ctx->ts_info.tile_rows;
-
-    ctx->tile_to_slice_map[0] = 0;
-    /* alloc tile information */
-    size = sizeof(XEVE_TILE) * f_tile;
-    ctx->tile = xeve_malloc(size);
-    xeve_assert_rv(ctx->tile, XEVE_ERR_OUT_OF_MEMORY);
-    xeve_mset(ctx->tile, 0, size);
-
-    /* update tile information - Tile width, height, First ctb address */
-    tidx = 0;
-    tile = &ctx->tile[tidx];
-    tile->w_ctb = ctx->w_lcu;
-    tile->h_ctb = ctx->h_lcu;
-    tile->f_ctb = tile->w_ctb * tile->h_ctb;
-
-    return XEVE_OK;
-}
-
-int xeve_ready(XEVE_CTX* ctx)
-{
-    XEVE_CORE* core = NULL;
-    int          w, h, ret, i, f_blk;
-    s32          size;
-    XEVE_FCST* fcst = &ctx->fcst;
-
-    xeve_assert(ctx);
-    if (ctx->core[0] == NULL)
-    {
-
-        /* set various value */
-        for (int i = 0; i < ctx->param.threads; i++)
-        {
-            core = xeve_core_alloc(ctx->param.chroma_format_idc);
-            xeve_assert_gv(core != NULL, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-            ctx->core[i] = core;
-        }
-    }
-
-    xeve_init_bits_est();
-
-    if (ctx->w == 0)
-    {
-        w = ctx->w = XEVE_ALIGN_VAL(ctx->param.w, 8);
-        h = ctx->h = XEVE_ALIGN_VAL(ctx->param.h, 8);
-        ctx->f = w * h;
-        if ((ctx->w != ctx->param.w) || (ctx->h != ctx->param.h))
-        {
-            ctx->param.picture_cropping_flag = 1;
-            ctx->param.picture_crop_left_offset = 0;
-            ctx->param.picture_crop_right_offset = (ctx->w - ctx->param.w + 1) >> 1;
-            ctx->param.picture_crop_top_offset = 0;
-            ctx->param.picture_crop_bottom_offset = (ctx->h - ctx->param.h + 1) >> 1;
-        }
-
-        ctx->max_cuwh = 64;
-        ctx->min_cuwh = 1 << 2;
-        ctx->log2_min_cuwh = 2;
-
-        ctx->log2_max_cuwh = XEVE_LOG2(ctx->max_cuwh);
-        ctx->max_cud = ctx->log2_max_cuwh - MIN_CU_LOG2;
-        ctx->w_lcu = (w + ctx->max_cuwh - 1) >> ctx->log2_max_cuwh;
-        ctx->h_lcu = (h + ctx->max_cuwh - 1) >> ctx->log2_max_cuwh;
-        ctx->f_lcu = ctx->w_lcu * ctx->h_lcu;
-        ctx->w_scu = (w + ((1 << MIN_CU_LOG2) - 1)) >> MIN_CU_LOG2;
-        ctx->h_scu = (h + ((1 << MIN_CU_LOG2) - 1)) >> MIN_CU_LOG2;
-        ctx->f_scu = ctx->w_scu * ctx->h_scu;
-        ctx->log2_culine = ctx->log2_max_cuwh - MIN_CU_LOG2;
-        ctx->log2_cudim = ctx->log2_culine << 1;
-    }
-
-    if (ctx->param.rc_type != 0 || ctx->param.lookahead != 0 || ctx->param.use_fcst != 0)
-    {
-        xeve_rc_create(ctx);
-    }
-    else
-    {
-        ctx->rc = NULL;
-        ctx->rcore = NULL;
-        ctx->qp = ctx->param.qp;
-    }
-
-    //initialize the threads to NULL
-    for (int i = 0; i < XEVE_MAX_THREADS; i++)
-    {
-        ctx->thread_pool[i] = 0;
-    }
-
-    //get the context synchronization handle
-    ctx->sync_block = get_synchronized_object();
-    xeve_assert_gv(ctx->sync_block != NULL, ret, XEVE_ERR_UNKNOWN, ERR);
-
-    if (ctx->param.threads >= 1)
-    {
-        ctx->tc = xeve_malloc(sizeof(THREAD_CONTROLLER));
-        init_thread_controller(ctx->tc, ctx->param.threads);
-        for (int i = 0; i < ctx->param.threads; i++)
-        {
-            ctx->thread_pool[i] = ctx->tc->create(ctx->tc, i);
-            xeve_assert_gv(ctx->thread_pool[i] != NULL, ret, XEVE_ERR_UNKNOWN, ERR);
-        }
-    }
-
-    size = ctx->f_lcu * sizeof(int);
-    ctx->sync_flag = (volatile s32 *)xeve_malloc(size);
-    xeve_assert_gv( ctx->sync_flag, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-    for (int i = 0; i < (int)ctx->f_lcu; i++)
-    {
-        ctx->sync_flag[i] = 0;
-    }
-
-    /*  allocate CU data map*/
-    if (ctx->map_cu_data == NULL)
-    {
-        size = sizeof(XEVE_CU_DATA) * ctx->f_lcu;
-        ctx->map_cu_data = (XEVE_CU_DATA*)xeve_malloc_fast(size);
-        xeve_assert_gv(ctx->map_cu_data, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-        xeve_mset_x64a(ctx->map_cu_data, 0, size);
-
-        for (i = 0; i < (int)ctx->f_lcu; i++)
-        {
-            xeve_create_cu_data(ctx->map_cu_data + i, ctx->log2_max_cuwh - MIN_CU_LOG2, ctx->log2_max_cuwh - MIN_CU_LOG2, ctx->param.chroma_format_idc);
-        }
-    }
-
-    /* allocate maps */
-    if (ctx->map_scu == NULL)
-    {
-        size = sizeof(u32) * ctx->f_scu;
-        ctx->map_scu = xeve_malloc_fast(size);
-        xeve_assert_gv(ctx->map_scu, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-        xeve_mset_x64a(ctx->map_scu, 0, size);
-    }
-
-    if (ctx->map_ipm == NULL)
-    {
-        size = sizeof(s8) * ctx->f_scu;
-        ctx->map_ipm = xeve_malloc_fast(size);
-        xeve_assert_gv(ctx->map_ipm, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-        xeve_mset(ctx->map_ipm, -1, size);
-    }
-
-    size = sizeof(s8) * ctx->f_scu;
-    ctx->map_depth = xeve_malloc_fast(size);
-    xeve_assert_gv(ctx->map_depth, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-    xeve_mset(ctx->map_depth, -1, size);
-
-
-    if (ctx->map_cu_mode == NULL)
-    {
-        size = sizeof(u32) * ctx->f_scu;
-        ctx->map_cu_mode = xeve_malloc_fast(size);
-        xeve_assert_gv(ctx->map_cu_mode, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-        xeve_mset_x64a(ctx->map_cu_mode, 0, size);
-    }
-
-    /* initialize reference picture manager */
-    ctx->pa.fn_alloc = xeve_pic_alloc;
-    ctx->pa.fn_free = xeve_pic_free;
-    ctx->pa.w = ctx->w;
-    ctx->pa.h = ctx->h;
-    ctx->pa.pad_l = PIC_PAD_SIZE_L;
-    ctx->pa.pad_c = PIC_PAD_SIZE_L >> ctx->param.cs_h_shift;
-    ctx->pa.bit_depth = ctx->param.codec_bit_depth;
-    ctx->pic_cnt = 0;
-    ctx->pic_icnt = -1;
-    ctx->poc.poc_val = 0;
-    ctx->pa.chroma_format_idc = ctx->param.chroma_format_idc;
-
-    ret = xeve_picman_init(&ctx->rpm, MAX_PB_SIZE, XEVE_MAX_NUM_REF_PICS, &ctx->pa);
-    xeve_assert_g(XEVE_SUCCEEDED(ret), ERR);
-
-    if (ctx->param.gop_size == 1 && ctx->param.keyint != 1) //LD case
-    {
-        ctx->pico_max_cnt = 2;
-    }
-    else //RA case
-    {
-        ctx->pico_max_cnt = 1 + ((ctx->param.gop_size) << 1);
-    }
-
-    if (ctx->param.bframes)
-    {
-        ctx->frm_rnum = XEVE_MAX(ctx->param.bframes + 1, ctx->param.lookahead);
-    }
-    else
-    {
-        ctx->frm_rnum = 0;
-    }
-
-    ctx->qp = ctx->param.qp;
-    if (ctx->param.use_fcst)
-    {
-        fcst->log2_fcst_blk_spic = 4; /* 16x16 in half image*/
-        fcst->w_blk = (ctx->w/2 + (((1 << (fcst->log2_fcst_blk_spic + 1)) - 1))) >> (fcst->log2_fcst_blk_spic + 1);
-        fcst->h_blk = (ctx->h/2 + (((1 << (fcst->log2_fcst_blk_spic + 1)) - 1))) >> (fcst->log2_fcst_blk_spic + 1);
-        fcst->f_blk = fcst->w_blk * fcst->h_blk;
-    }
-
-    for (i = 0; i < ctx->pico_max_cnt; i++)
-    {
-        ctx->pico_buf[i] = (XEVE_PICO*)xeve_malloc(sizeof(XEVE_PICO));
-        xeve_assert_gv(ctx->pico_buf[i], ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-        xeve_mset(ctx->pico_buf[i], 0, sizeof(XEVE_PICO));
-
-        XEVE_PICO *pico;
-        pico = ctx->pico_buf[i];
-
-        pico->spic = xeve_alloc_spic_l(ctx->w, ctx->h);
-        xeve_assert_g(pico->spic != NULL, ERR);
-
-        if (ctx->param.use_fcst)
-        {
-            ctx->pico_buf[i]->spic = xeve_alloc_spic_l(ctx->w, ctx->h);
-
-            f_blk = ctx->fcst.f_blk;
-            size = sizeof(u8) * f_blk;
-            ctx->pico_buf[i]->sinfo.map_pdir = xeve_malloc(size);
-            xeve_assert_gv(ctx->pico_buf[i]->sinfo.map_pdir, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-
-            size = sizeof(u8) * f_blk;
-            ctx->pico_buf[i]->sinfo.map_pdir_bi = xeve_malloc(size);
-            xeve_assert_gv(ctx->pico_buf[i]->sinfo.map_pdir_bi, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-
-            size = sizeof(s16) * f_blk * PRED_BI * MV_D;
-            ctx->pico_buf[i]->sinfo.map_mv = xeve_malloc(size);
-            xeve_assert_gv(ctx->pico_buf[i]->sinfo.map_mv, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-
-            size = sizeof(s16) * f_blk * PRED_BI * MV_D;
-            ctx->pico_buf[i]->sinfo.map_mv_bi = xeve_malloc(size);
-            xeve_assert_gv(ctx->pico_buf[i]->sinfo.map_mv_bi, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-
-            size = sizeof(s16) * f_blk * PRED_BI * MV_D;
-            ctx->pico_buf[i]->sinfo.map_mv_pga = xeve_malloc(size);
-            xeve_assert_gv(ctx->pico_buf[i]->sinfo.map_mv_pga, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-
-            size = sizeof(s32) * f_blk * 4;
-            ctx->pico_buf[i]->sinfo.map_uni_lcost = xeve_malloc(size);
-            xeve_assert_gv(ctx->pico_buf[i]->sinfo.map_uni_lcost, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-
-            size = sizeof(s32) * f_blk;
-            ctx->pico_buf[i]->sinfo.map_bi_lcost = xeve_malloc(size);
-            xeve_assert_gv(ctx->pico_buf[i]->sinfo.map_bi_lcost, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-
-            size = sizeof(s32) * f_blk;
-            ctx->pico_buf[i]->sinfo.map_qp_blk = xeve_malloc(size);
-            xeve_assert_gv(ctx->pico_buf[i]->sinfo.map_qp_blk, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-
-            size = sizeof(s8) * ctx->f_scu;
-            ctx->pico_buf[i]->sinfo.map_qp_scu = xeve_malloc_fast(size);
-            xeve_assert_gv(ctx->pico_buf[i]->sinfo.map_qp_scu, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-
-            size = sizeof(u16) * f_blk;
-            ctx->pico_buf[i]->sinfo.transfer_cost = xeve_malloc(size);
-            xeve_assert_gv(ctx->pico_buf[i]->sinfo.transfer_cost, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-        }
-    }
-
-    /* alloc tile index map in SCU unit */
-    if (ctx->map_tidx == NULL)
-    {
-        size = sizeof(u8) * ctx->f_scu;
-        ctx->map_tidx = (u8*)xeve_malloc_fast(size);
-        xeve_assert_gv(ctx->map_tidx, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-        xeve_mset_x64a(ctx->map_tidx, 0, size);
-    }
-
-    if (ctx->tile == NULL)
-    {
-        ret = ctx->fn_set_tile_info(ctx);
-        if (ret != XEVE_OK)
-        {
-            goto ERR;
-        }
-    }
-
-    ctx->sh_array = (XEVE_SH*)xeve_malloc(sizeof(XEVE_SH) * ctx->ts_info.num_slice_in_pic);
-    xeve_assert_gv(ctx->sh_array, ret, XEVE_ERR_OUT_OF_MEMORY, ERR);
-    xeve_mset(ctx->sh_array, 0, sizeof(XEVE_SH) * ctx->ts_info.num_slice_in_pic);
-    ctx->sh = &ctx->sh_array[0];
-
-    return XEVE_OK;
-ERR:
-    for (i = 0; i < (int)ctx->f_lcu; i++)
-    {
-        xeve_delete_cu_data(ctx->map_cu_data + i, ctx->log2_max_cuwh - MIN_CU_LOG2, ctx->log2_max_cuwh - MIN_CU_LOG2);
-    }
-
-    xeve_mfree_fast(ctx->map_cu_data);
-    xeve_mfree_fast(ctx->map_ipm);
-    xeve_mfree_fast(ctx->map_depth);
-    xeve_mfree_fast(ctx->map_cu_mode);
-    xeve_mfree_fast(ctx->sh_array);
-
-    //free the threadpool and created thread if any
-    if (ctx->sync_block)
-    {
-        release_synchornized_object(&ctx->sync_block);
-    }
-
-    if (ctx->param.threads >= 1)
-    {
-        if (ctx->tc)
-        {
-            //thread controller instance is present
-            //terminate the created thread
-            for (int i = 0; i < ctx->param.threads; i++)
-            {
-                if (ctx->thread_pool[i])
-                {
-                    //valid thread instance
-                    ctx->tc->release(&ctx->thread_pool[i]);
-                }
-            }
-            //dinitialize the tc
-            dinit_thread_controller(ctx->tc);
-            xeve_mfree_fast(ctx->tc);
-            ctx->tc = 0;
-        }
-    }
-
-    xeve_mfree_fast((void*)ctx->sync_flag);
-
-    if (ctx->param.use_fcst)
-    {
-        for (i = 0; i < ctx->pico_max_cnt; i++)
-        {
-            xeve_picbuf_rc_free(ctx->pico_buf[i]->spic);
-            xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_pdir);
-            xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_pdir_bi);
-            xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_mv);
-            xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_mv_bi);
-            xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_mv_pga);
-            xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_uni_lcost);
-            xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_bi_lcost);
-            xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_qp_blk);
-            xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_qp_scu);
-            xeve_mfree_fast(ctx->pico_buf[i]->sinfo.transfer_cost);
-            xeve_mfree_fast(ctx->pico_buf[i]);
-        }
-    }
-
-    if (core)
-    {
-        xeve_core_free(core);
-    }
-
-    if (ctx->param.rc_type != 0 || ctx->param.lookahead != 0)
-    {
-        xeve_rc_delete(ctx);
-    }
-
-    return ret;
-}
-
-void xeve_flush(XEVE_CTX * ctx)
-{
-    int i;
-    xeve_assert(ctx);
-
-    xeve_mfree_fast(ctx->map_scu);
-    for(i = 0; i < (int)ctx->f_lcu; i++)
-    {
-        xeve_delete_cu_data(ctx->map_cu_data + i, ctx->log2_max_cuwh - MIN_CU_LOG2, ctx->log2_max_cuwh - MIN_CU_LOG2);
-    }
-    xeve_mfree_fast(ctx->map_cu_data);
-    xeve_mfree_fast(ctx->map_ipm);
-    xeve_mfree_fast(ctx->map_depth);
-    xeve_mfree_fast(ctx->sh_array);
-
-    //release the sync block
-    if (ctx->sync_block)
-    {
-        release_synchornized_object(&ctx->sync_block);
-    }
-
-    //Release thread pool controller and created threads
-    if (ctx->param.threads >= 1)
-    {
-        if(ctx->tc)
-        {
-            //thread controller instance is present
-            //terminate the created thread
-            for (int i = 0; i < ctx->param.threads; i++)
-            {
-                if(ctx->thread_pool[i])
-                {
-                    //valid thread instance
-                    ctx->tc->release(&ctx->thread_pool[i]);
-                }
-            }
-            //dinitialize the tc
-            dinit_thread_controller(ctx->tc);
-            xeve_mfree_fast(ctx->tc);
-            ctx->tc = 0;
-        }
-    }
-
-        xeve_mfree_fast((void*) ctx->sync_flag);
-
-    xeve_mfree_fast(ctx->map_cu_mode);
-    xeve_picbuf_free(ctx->pic_dbk);
-    xeve_picman_deinit(&ctx->rpm);
-
-    for (int i = 0; i < ctx->param.threads; i++)
-    {
-        xeve_core_free(ctx->core[i]);
-    }
-
-    for(i = 0; i < ctx->pico_max_cnt; i++)
-    {
-        xeve_picbuf_rc_free(ctx->pico_buf[i]->spic);
-
-        xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_pdir);
-        xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_pdir_bi);
-        xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_mv);
-        xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_mv_bi);
-        xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_mv_pga);
-        xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_uni_lcost);
-        xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_bi_lcost);
-        xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_qp_blk);
-        xeve_mfree_fast(ctx->pico_buf[i]->sinfo.map_qp_scu);
-        xeve_mfree_fast(ctx->pico_buf[i]->sinfo.transfer_cost);
-        xeve_mfree_fast(ctx->pico_buf[i]);
-    }
-
-    for(i = 0; i < XEVE_MAX_INBUF_CNT; i++)
-    {
-        if(ctx->inbuf[i]) ctx->inbuf[i]->release(ctx->inbuf[i]);
-    }
-
-    if (ctx->param.rc_type != 0 ||  ctx->param.aq_mode !=0)
-    {
-        xeve_rc_delete(ctx);
-    }
-}
-
-int xeve_picbuf_get_inbuf(XEVE_CTX * ctx, XEVE_IMGB ** imgb)
-{
-    int i, opt, align[XEVE_IMGB_MAX_PLANE], pad[XEVE_IMGB_MAX_PLANE];
-
-    for(i = 0; i < XEVE_MAX_INBUF_CNT; i++)
-    {
-        if(ctx->inbuf[i] == NULL)
-        {
-            opt = XEVE_IMGB_OPT_NONE;
-
-            /* set align value*/
-            align[0] = MIN_CU_SIZE;
-            align[1] = MIN_CU_SIZE;
-            align[2] = MIN_CU_SIZE;
-
-            /* no padding */
-            pad[0] = 0;
-            pad[1] = 0;
-            pad[2] = 0;
-
-            int cs = ctx->param.chroma_format_idc == 0 ? XEVE_CS_YCBCR400_10LE : (ctx->param.chroma_format_idc == 1 ? XEVE_CS_YCBCR420_10LE :
-                (ctx->param.chroma_format_idc == 2 ? XEVE_CS_YCBCR422_10LE : XEVE_CS_YCBCR444_10LE));
-            *imgb = xeve_imgb_create(ctx->w, ctx->h, cs, opt, pad, align);
-            xeve_assert_rv(*imgb != NULL, XEVE_ERR_OUT_OF_MEMORY);
-
-            ctx->inbuf[i] = *imgb;
-
-            (*imgb)->addref(*imgb);
-            return XEVE_OK;
-        }
-        else if(ctx->inbuf[i]->getref(ctx->inbuf[i]) == 1)
-        {
-            *imgb = ctx->inbuf[i];
-
-            (*imgb)->addref(*imgb);
-            return XEVE_OK;
-        }
-    }
-
-    return XEVE_ERR_UNEXPECTED;
-}
-
-int xeve_header(XEVE_CTX * ctx)
-{
-    int ret = XEVE_OK;
-
-    /* encode parameter sets */
-    if (ctx->pic_cnt == 0 || (ctx->slice_type == SLICE_I && ctx->param.closed_gop)) /* if nalu_type is IDR */
-    {
-        ret = ctx->fn_encode_sps(ctx);
-        xeve_assert_rv(ret == XEVE_OK, ret);
-
-        ret = ctx->fn_encode_pps(ctx);
-        xeve_assert_rv(ret == XEVE_OK, ret);
-    }
-
-    return ret;
 }
 
 void xeve_update_core_loc_param(XEVE_CTX * ctx, XEVE_CORE * core)
