+++ src_base/xeve_util.h
@@ -196,24 +196,7 @@ int  xeve_check_cpu_info();
     (ctx) = (XEVE_CTX *)id; \
     xeve_assert_rv((ctx)->magic == XEVE_MAGIC_CODE, (ret));
 
-int  xeve_encode_sps(XEVE_CTX * ctx);
-int  xeve_encode_pps(XEVE_CTX * ctx);
-
-XEVE_CTX * xeve_ctx_alloc(void);
-void xeve_ctx_free(XEVE_CTX * ctx);
-XEVE_CORE * xeve_core_alloc(int chroma_format_idc);
-void xeve_core_free(XEVE_CORE * core);
 void xeve_copy_chroma_qp_mapping_params(XEVE_CHROMA_TABLE *dst, XEVE_CHROMA_TABLE *src);
-int  xeve_set_init_param(XEVE_CTX * ctx, XEVE_PARAM * param);
-int  xeve_pic_finish(XEVE_CTX *ctx, XEVE_BITB *bitb, XEVE_STAT *stat);
-void xeve_set_nalu(XEVE_NALU * nalu, int nalu_type, int nuh_temporal_id);
-void xeve_set_vui(XEVE_CTX * ctx, XEVE_VUI * vui);
-void xeve_set_sps(XEVE_CTX * ctx, XEVE_SPS * sps);
-void xeve_set_pps(XEVE_CTX * ctx, XEVE_PPS * pps);
-int  xeve_set_active_pps_info(XEVE_CTX * ctx);
-void xeve_set_sh(XEVE_CTX *ctx, XEVE_SH *sh);
-int  xeve_set_tile_info(XEVE_CTX * ctx);
-int  xeve_header(XEVE_CTX * ctx);
 void xeve_update_core_loc_param(XEVE_CTX * ctx, XEVE_CORE * core);
 void xeve_update_core_loc_param_mt(XEVE_CTX * ctx, XEVE_CORE * core);
 int  xeve_mt_get_next_ctu_num(XEVE_CTX * ctx, XEVE_CORE * core, int skip_ctb_line_cnt);
